IKSWP-76402: ASoC: sm6150: create one new initial dailink for doha
WSA is not present on Doha product, remove some component which is not
present on board. This is helpful to create sound card successfully.
I think this is enough for debuging wcd9370 on doha board.

Tacna and cs35l41 is to be added.

Change-Id: Iad3f78d14b5babc3663232be75a66216660b4ef6
Signed-off-by: huangys <huangys@motorola.com>
Reviewed-on: https://gerrit.mot.com/1338337
SLTApproved: Slta Waiver
SME-Granted: SME Approvals Granted
Tested-by: Jira Key
Submit-Approved: Jira Key